✨ Added win rate and shards
